Order rows using column values

arrange() orders the rows of a data frame by the values of selected columns.

Unlike other dplyr verbs, arrange() largely ignores grouping; you need to explicitly mention grouping variables (or use .by_group = TRUE) in order to group by them, and functions of variables are evaluated once per data frame, not once per group.

Usage

# S3 method for class 'codingsystem'
arrange(.data, ..., .by_group = FALSE)

Arguments

.data

A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr). See Methods, below, for more details.

...

<data-masking> Variables, or functions of variables. Use desc() to sort a variable in descending order.

.by_group

If TRUE, will sort first by grouping variable. Applies to grouped data frames only.

Value

An object of the same type as .data. The output has the following properties:

All rows appear in the output, but (usually) in a different place.

Columns are not modified.

Groups are not modified.

Data frame attributes are preserved.

Details

Missing values

Unlike base sorting with sort(), NA are:

always sorted to the end for local data, even when wrapped with desc().

treated differently for remote data, depending on the backend.

Methods

This function is a generic, which means that packages can provide implementations (methods) for other classes. See the documentation of individual methods for extra arguments and differences in behaviour.

The following methods are currently available in loaded packages: no methods found .

See also

Examples

arrange(mtcars, cyl, disp)
#> Error in arrange(mtcars, cyl, disp): could not find function "arrange"
arrange(mtcars, desc(disp))
#> Error in arrange(mtcars, desc(disp)): could not find function "arrange"

# grouped arrange ignores groups
by_cyl <- mtcars %>% group_by(cyl)
#> Error in mtcars %>% group_by(cyl): could not find function "%>%"
by_cyl %>% arrange(desc(wt))
#> Error in by_cyl %>% arrange(desc(wt)): could not find function "%>%"
# Unless you specifically ask:
by_cyl %>% arrange(desc(wt), .by_group = TRUE)
#> Error in by_cyl %>% arrange(desc(wt), .by_group = TRUE): could not find function "%>%"

# use embracing when wrapping in a function;
# see ?rlang::args_data_masking for more details
tidy_eval_arrange <- function(.data, var) {
  .data %>%
    arrange({{ var }})
}
tidy_eval_arrange(mtcars, mpg)
#> Error in .data %>% arrange({    {        var    }}): could not find function "%>%"

# Use `across()` or `pick()` to select columns with tidy-select
iris %>% arrange(pick(starts_with("Sepal")))
#> Error in iris %>% arrange(pick(starts_with("Sepal"))): could not find function "%>%"
iris %>% arrange(across(starts_with("Sepal"), desc))
#> Error in iris %>% arrange(across(starts_with("Sepal"), desc)): could not find function "%>%"